requirements.txt
pillow==7.1.2
numpy==1.18.1
opencv-python==4.2.0.34
imutils==0.5.3
pytesseract==0.3.4
flask==1.1.2
gunicorn==20.0.4
mrz==0.6.2